Move KubeCardsPlay component.
