+++ src/kube-cards-website/src/KubeCardsContent.tsx
@@ -5,7 +5,7 @@ import { withStyles, WithStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import { Typography } from '@material-ui/core';
 
-import KubeCardsPlay from './KubeCardsPlay';
+import KubeCardsPlay from './components/play/KubeCardsPlay';
 
 const styles = (theme: any) => ({
     root: {
